Perfil adaptado con localstorage - datos actualizar , foto de perfil , navbar con foto de perfil y actualizacion de datos de mascotas
